efl: Make stable single-valued getters consistent.
Summary:
Instead of a getter with an explicit return type, change it to be a
single-valued property.

The eolian C generator takes care of making this single value the actual
return value of the C function.

This also makes these properties able to be reflected on.

The stack properties returns just a pointer and not a new ref, so no
@move needed.

Beta properties will be handled in a future commit.

Depends on D10601

Reviewers: segfaultxavi, bu5hm4n, q66, cedric

Reviewed By: segfaultxavi

Subscribers: #reviewers, #committers, brunobelo, felipealmeida

Tags: #efl

Differential Revision: https://phab.enlightenment.org/D10602