csharp: FunctionPtr wrappers are not IDisposable
Summary:
They don't actually store unmanaged memory and we already rely on the GC
to free them (not calling Dispose() directly through the managedCb
closure, so we don't need IDisposable.

Just making sure we release the Function Pointer data in the main loop
as the GC runs in a different thread.

Ref T8423

Depends on D10672

Reviewers: brunobelo, felipealmeida, YOhoho

Reviewed By: brunobelo

Subscribers: cedric, #reviewers, #committers

Tags: #efl

Maniphest Tasks: T8423

Differential Revision: https://phab.enlightenment.org/D10673