tooling: add configuration file for lgtm.com service
It seems they run this on many Github projects automatically now:
https://lgtm.com/projects/g/Enlightenment/efl

From my playing around it had 970 "recommendations" to fix our FIXME
tags. Filtering out this noise by using this config file to exclude the
query.

It also seems to pick up the python scripts as Python2 and has wrong
reporting due to this as well.

Signed-off-by: Stefan Schmidt <s.schmidt@samsung.com>
Reviewed-by: Marcel Hollerbach <mail@marcel-hollerbach.de>
Differential Revision: https://phab.enlightenment.org/D10699