WIP csharp: Fix passing acessor with ownership
When passing an owned acessor from a converted collection we need a way
to unpin the passed data when the accessor is freed.

This commits adds a thin wrapper around the CArray accessor that unpins
the data when freed.

NOTE: Needs to update the test after D10878 commented out the accessor
indexer