ci: re-enable cxx bindings for CI builds again
While cxx bindings are enabled by default we had them turned off on the
CI builds due to the amount of memory they used in the examples
building/linking. This often had lead to the case where the Travis build
failed due too little resources on the Travis nodes for it.

I have not seen this stalls during linking on my machine for a while now
and my tests on Travis do also not show this. Time to enable them again
and if the problem comes back we can revert.

Signed-off-by: Stefan Schmidt <s.schmidt@samsung.com>
Reviewed-by: Marcel Hollerbach <mail@marcel-hollerbach.de>
Differential Revision: https://phab.enlightenment.org/D10822