ci: create a daily build to run with coverage enabled and upload to codecov
Summary:
For this we bring back a native Linux build on the Travis with Ubuntu
bionic. This allows use easy integration with codecov as Travis and
Codecov have figured out all details.
This means we need to be a bit more careful with our $DIST checks as
empty no longer means osx. It could as well be the Linux job.

Covergae reports will show up here:
https://codecov.io/gh/Enlightenment/efl

We are starting with 36%. Time to improve. :-)

Fixes: T7910

Reviewers: bu5hm4n, zmike

Reviewed By: zmike

Subscribers: cedric

Tags: #efl

Differential Revision: https://phab.enlightenment.org/D10867