efl_ui_spotlight_manager_stack: do not keep final state
Summary:
keeping the final state here results in the fact that the map is not
getting reseted. Which is a problem, as the element would not be
displayed correctly after that.

Final state here means, that the last configuration of Efl.Gfx.Mapping will be kept, and it will *not* be resetted.

Reviewers: zmike, cedric, segfaultxavi

Reviewed By: zmike

Subscribers: #reviewers, #committers

Tags: #efl

Differential Revision: https://phab.enlightenment.org/D10999