ecore-x add root barrier setup func to fix xserver mouse containment
xserver stopped containing mouse to screen bounds a while back... this
si broken. so enforce this policy with an api that take a list of
screen rects (relative to root) and makes those the barrier bounds so
that mouse doesn't go out of the screen anymore. new api to enable
this fix in e.