ci: travis: add build on ARM64 arch now provided by Travis
Summary:
Cheap and easy way for us to have basic compile testing on ARM64. Travis
is also offering ppc64le and s390x for builds, but on ppc64le we get a
internal compoler error (yay!) on c++ examples and on s390x we have no
luajit. So, these two have to wait a bit more.

Reviewers: bu5hm4n, zmike

Reviewed By: zmike

Subscribers: cedric, #reviewers, #committers

Tags: #efl

Differential Revision: https://phab.enlightenment.org/D11060