efl_ui_progressbar: fix theming of progressbar
There was code in progressbar that parsed its own theme name, applied a
change, reassembed it, and set it again. However, when we simply stick
to the theme apply code path that we have in efl.ui.layout_base, we can
simply set the horizontal / vertical part to be the element, the rest is
the class. This safes a lot of code, additionally, the setting of the
spacer had to be moved to the theme_apply method, as this would not work
otherwise.

ref T8501

Reviewed-by: Mike Blumenkrantz <michael.blumenkrantz@gmail.com>
Differential Revision: https://phab.enlightenment.org/D11076