ler-plugin: initial gcc plugin harness to be used in our build
Ideas and some code from https://github.com/mesonbuild/meson/issues/5090

This is just an example plugin that cold be used as a starting point.
Right now we find efl_super() with GIMPLE and add a efl_super2 call when
it was defined and in code before to allow linking.