webp - promote to default on
i just noticed a pattern... we recommend in our sample confs ... to
not disable webp. why keep doing that and why not just make it a dep
on by default you need to explicitly disable? make lives easier and
less complex. it was a good exercise to write these as it points this
out... :)