TH - go back to nearest smoothing fill for shadows
there's a rounding problem... and a single part with image with hole
in it just are going to always fail in lots of cases. i need to
rethink the shadows - probably break up into 4 or 8 parts with
clipping to solve this... so for now this will do until i get to that.